use tracking memory resource in memory resource manager

+++ src/lib/utils/memory_resource_manager.hpp
@@ -14,7 +14,7 @@ class MemoryResourceManager : public Noncopyable {
 
   // TODO comment
   const std::unordered_map<std::string, std::shared_ptr<TrackingMemoryResource>>& memory_resources() const;
-  const std::unordered_map<std::string, int64_t> get_current_memory_usage() const;
+  const std::unordered_map<std::string, size_t> get_current_memory_usage() const;
   std::shared_ptr<TrackingMemoryResource> get_memory_resource(const std::string& purpose);
 
  protected:
